dialog example: Separate provider specific code into modules
